Clean DAG: each layer depends only on layers below it
- dom-transform.js → utils/ (DOM is not pure core)
- config.js: remove window fallback, adapter calls initConfig() explicitly
- export-renderers.js split → renderers/latex.js, typst.js, markdown.js, index.js
- core/yaml-components.js deleted: assembly moved to src/index.js (no core→renderers edge)
- COMPONENT_REGISTRY removed from markdown-utils.js (duplicated defaultComponents keys)
- Astro adapter: inject components map as param, drop global side-effect loading

DAG is now: core ← renderers ← index.js ← adapters/
